Move insert_content to a tool file (#2095)
